Configure secret credentials
