+++ .gitignore
@@ -37,4 +37,4 @@ captures/
 # IntelliJ
 *.iml
 .idea/
-/Examples/src/main/java/com/pubnub/crc/chat_examples_java/util/Keys.java
+/Examples/gradle.properties
