+++ .travis.yml
@@ -25,13 +25,16 @@ matrix:
 
 before_install:
   - chmod +x gradlew
+  - echo PUB_KEY="\"${PUB_KEY}\"" >> gradle.properties
+  - echo SUB_KEY="\"${SUB_KEY}\"" >> gradle.properties
+  - cat gradle.properties
   - curl -sL https://deb.nodesource.com/setup_9.x | sudo -E bash -
   - sudo apt-get install nodejs
   - sudo apt-get install build-essential
 
 install:
   - npm config set strict-ssl false
-  - npm install chalk fancy-log simple-git   
+  - npm install chalk fancy-log simple-git
 
 before_script:
   - git config --global user.name "${GH_NAME}"
