+++ Examples/src/main/AndroidManifest.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.pubnub.crc.chat_examples_java">
+    package="com.pubnub.crc.sample">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".App"
+        android:name="com.pubnub.crc.sample.App"
         android:allowBackup="false"
         android:appComponentFactory="androidx.core.app.CoreComponentFactory"
         android:icon="@mipmap/ic_launcher"
@@ -18,7 +18,7 @@
         tools:replace="android:appComponentFactory"
         tools:targetApi="p">
         <activity
-            android:name=".SplashActivity"
+            android:name="com.pubnub.crc.sample.SplashActivity"
             android:theme="@style/SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -27,7 +27,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name="com.pubnub.crc.sample.MainActivity"
             android:theme="@style/AppTheme" />
     </application>
