--- Examples/src/main/java/com/pubnub/crc/chat_examples_java/util/PNFragment.java
@@ -1,9 +1,0 @@
-package com.pubnub.crc.chat_examples_java.util;
-
-import com.pubnub.api.callbacks.SubscribeCallback;
-
-public interface PNFragment {
-
-    SubscribeCallback provideListener();
-
-}
