--- Examples/src/main/java/com/pubnub/crc/chat_examples_java/util/PNGlideModule.java
@@ -1,9 +1,0 @@
-package com.pubnub.crc.chat_examples_java.util;
-
-import com.bumptech.glide.annotation.GlideModule;
-import com.bumptech.glide.module.AppGlideModule;
-
-@GlideModule
-public class PNGlideModule extends AppGlideModule {
-
-}
