+++ Examples/src/main/java/com/pubnub/crc/sample/ParentActivity.java
@@ -1,0 +1,38 @@
+package com.pubnub.crc.sample;
+
+import android.os.Bundle;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+import butterknife.ButterKnife;
+
+public abstract class ParentActivity extends AppCompatActivity {
+
+    protected abstract int provideLayoutResourceId();
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(provideLayoutResourceId());
+        ButterKnife.bind(this);
+    }
+
+    protected void addFragment(Fragment fragment) {
+        FragmentManager manager = getSupportFragmentManager();
+        FragmentTransaction ft = manager.beginTransaction();
+        setupAnimations(ft);
+        ft.replace(R.id.container, fragment);
+        ft.addToBackStack(fragment.getClass().getSimpleName());
+        ft.commit();
+    }
+
+    private void setupAnimations(FragmentTransaction ft) {
+        /*ft.setCustomAnimations(R.anim.enter_from_right, R.anim.exit_to_left, R.anim.enter_from_left, R.anim
+                .exit_to_right);*/
+        ft.setCustomAnimations(R.anim.fade_in, R.anim.fade_out, R.anim.fade_in, R.anim.fade_out);
+    }
+
+}
