+++ Examples/src/main/java/com/pubnub/crc/sample/SplashActivity.java
@@ -1,0 +1,23 @@
+package com.pubnub.crc.sample;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.Handler;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
+public class SplashActivity extends AppCompatActivity implements Runnable {
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        new Handler().postDelayed(this, 200);
+    }
+
+    @Override
+    public void run() {
+        startActivity(new Intent(this, MainActivity.class));
+        finish();
+    }
+}
