+++ Examples/src/main/java/com/pubnub/crc/sample/util/PNFragment.java
@@ -1,0 +1,9 @@
+package com.pubnub.crc.sample.util;
+
+import com.pubnub.api.callbacks.SubscribeCallback;
+
+public interface PNFragment {
+
+    SubscribeCallback provideListener();
+
+}
