+++ Examples/src/main/java/com/pubnub/crc/sample/util/PNGlideModule.java
@@ -1,0 +1,9 @@
+package com.pubnub.crc.sample.util;
+
+import com.bumptech.glide.annotation.GlideModule;
+import com.bumptech.glide.module.AppGlideModule;
+
+@GlideModule
+public class PNGlideModule extends AppGlideModule {
+
+}
