+++ Examples/src/main/java/com/pubnub/crc/sample/util/ParentActivityImpl.java
@@ -1,0 +1,15 @@
+package com.pubnub.crc.sample.util;
+
+import com.pubnub.api.PubNub;
+
+import androidx.fragment.app.Fragment;
+
+public interface ParentActivityImpl {
+
+    PubNub getPubNub();
+
+    void setTitle(String title);
+
+    void addFragment(Fragment fragment);
+
+}
