+++ build.gradle
@@ -4,11 +4,11 @@ buildscript {
     repositories {
         google()
         jcenter()
-        
+
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.3.2'
-        
+
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -19,8 +19,20 @@ allprojects {
         google()
         jcenter()
     }
+
+    task checkstyle(type: Checkstyle) {
+        configFile new File(rootDir, "config/checkstyle/checkstyle.xml")
+        source 'src/test'
+        include '**/*.java'
+        exclude '**/gen/**'
+        exclude '**/R.java'
+        exclude '**/BuildConfig.java'
+        classpath = files()
+    }
 }
 
 task clean(type: Delete) {
     delete rootProject.buildDir
 }
+
+
